move spy_dulmage_mendelsohn to visualize module
